Increase event timeStamp resolution
Depends on https://github.com/lightpanda-io/libdom/pull/36

The spec says this should be a High Definition timestamp. But browsers avoid
that in order to avoid fingerprinting. By default, FireFox rounds to 2ms (which
is what this PR does).

Previously, the timestamp was seconds, so you'd think: isn't that better? Well,
it's pretty far off the spec and what browsers do, but more importantly, it
crashes our WPT test. If you look at `Event-timestamp-safe-resolution.html`
you'll see that it's trying to find the delta between two timestamps, in an
endless loop (without a loop of many iterations). With second-resolution, it
just takes too long (and crashes..memory).

+++ src/browser/events/event.zig
@@ -113,7 +113,7 @@ pub const Event = struct {
     // Even though this is supposed to to provide microsecond resolution, browser
     // return coarser values to protect against fingerprinting. libdom returns
     // seconds, which is good enough.
-    pub fn get_timeStamp(self: *parser.Event) !u32 {
+    pub fn get_timeStamp(self: *parser.Event) !u64 {
         return parser.eventTimestamp(self);
     }
